buffer: Better fix for alignment of multiline messages
This one makes it easier to see which lines are separate messages and
which are just a multiline message.

Inspired (actually shamelessly copy-pasted) by
https://github.com/wee-slack/wee-slack/commit/adb1029977169ac665041e2e95613c531808875e
that fixes the same issue in wee-slack.

Example before:

    20:20:00 <abcdefg> > In reply to @xyz:matrix.org
    20:20:00 <abcdefg> > Lorem ipsum dolor sit amet, consectetur adipiscing elit.
    20:20:00 <abcdefg> Sed accumsan nisi vel vestibulum hendrerit.
    20:21:00 <xyz> Cras a lacus libero.

Example after the fix:

    20:20:00 <abcdefg> > In reply to @xyz:matrix.org
    20:20:00 <       > > Lorem ipsum dolor sit amet, consectetur adipiscing elit.
    20:20:00 <       > Sed accumsan nisi vel vestibulum hendrerit.
    20:21:00 <xyz> Cras a lacus libero.

This unfortunately breaks prefix_same_nick for multiline messages again.
(Probably not a big deal. Certainly not for me.)